Separate hbs client loader and builder
In order to be able to require different modules at build time and in
the client side, we have to split the client loader and the builder.

This way, we require a 'handlebars-compiler' for the builder, so we can
set the 'handlebars' module to the runtime-only file at build time and
when using the built package.

One other benefit is the final bundle overhead will be smaller in size.

It appears that both files can't share any code, so we have to specify
the extension on each one.