Much simpler runtime only inclusion and example
Removed all the tricks I added to load the handlebars compiler when
needed. No more excludeShallow, map, handlebars-compiler etc. The
configuration is now much more simple to understand.

Introduced a new configuration option to locate the handlebars library
to be used by the builder.

Cleaned up the example, so there is only one showcase with comments to
let the user fine tune the configuration. This should avoid some
confusion.